# jaiswal-store/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jaiswal Store API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  discount: number;
  imageUrl?: string | null;
  finalPrice: number;
  savings: number;
}

export interface CreateProductBody {
  name: string;
  price: number;
  discount: number;
  imageUrl?: string | null;
}

export interface UpdateProductBody {
  name: string;
  price: number;
  discount: number;
  imageUrl?: string | null;
}

export interface ProductStats {
  totalProducts: number;
  onDiscount: number;
  avgDiscount: number;
  lowestPrice?: number | null;
  highestPrice?: number | null;
}

export interface ShopSettings {
  id: number;
  name: string;
  phone?: string | null;
  address?: string | null;
}

export interface UpdateShopSettingsBody {
  name: string;
  phone?: string | null;
  address?: string | null;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface LoginResponse {
  username: string;
  message: string;
}

export interface SessionInfo {
  isAdmin: boolean;
  username?: string | null;
}

export interface ChangeCredentialsBody {
  username: string;
  password: string;
}