# jaiswal-store/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jaiswal Store API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all products
 */
export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  price: zod.number(),
  discount: zod.number(),
  imageUrl: zod.string().nullish(),
  finalPrice: zod.number(),
  savings: zod.number(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Create a product (admin only)
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  price: zod.number(),
  discount: zod.number(),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Get a product by ID
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  price: zod.number(),
  discount: zod.number(),
  imageUrl: zod.string().nullish(),
  finalPrice: zod.number(),
  savings: zod.number(),
});

/**
 * @summary Update a product (admin only)
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProductBody = zod.object({
  name: zod.string(),
  price: zod.number(),
  discount: zod.number(),
  imageUrl: zod.string().nullish(),
});

export const UpdateProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  price: zod.number(),
  discount: zod.number(),
  imageUrl: zod.string().nullish(),
  finalPrice: zod.number(),
  savings: zod.number(),
});

/**
 * @summary Delete a product (admin only)
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get product statistics
 */
export const GetProductStatsResponse = zod.object({
  totalProducts: zod.number(),
  onDiscount: zod.number(),
  avgDiscount: zod.number(),
  lowestPrice: zod.number().nullish(),
  highestPrice: zod.number().nullish(),
});

/**
 * @summary Get shop settings
 */
export const GetShopSettingsResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string().nullish(),
  address: zod.string().nullish(),
});

/**
 * @summary Update shop settings (admin only)
 */
export const UpdateShopSettingsBody = zod.object({
  name: zod.string(),
  phone: zod.string().nullish(),
  address: zod.string().nullish(),
});

export const UpdateShopSettingsResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string().nullish(),
  address: zod.string().nullish(),
});

/**
 * @summary Admin login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  username: zod.string(),
  message: zod.string(),
});

/**
 * @summary Get current session info
 */
export const GetMeResponse = zod.object({
  isAdmin: zod.boolean(),
  username: zod.string().nullish(),
});

/**
 * @summary Change admin credentials (admin only)
 */
export const ChangeCredentialsBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});